pacman: add mingw-w64-clang-aarch64-asciidoc
asciidoctor is not available for clangarm64 yet as it
depends on ruby, which is also not available yet. Therefore
we use asciidoc. While there is an asciidoc executable
available in /usr/bin/asciidoc, it runs under x64 emulation
which is very slow. Let's install the native
mingw-w64-clang-aarch64-asciidoc which will significantly
speed up our build.

Signed-off-by: Dennis Ameling <dennis@dennisameling.com>