update-via-pacman: preserve `IgnorePkg` lines in `pacman.conf`
Git for Windows needs to prevent a couple of packages from being
upgraded, for now. To this end, the `IgnorePkg` line in `pacman.conf`
specifies the list of these packages.

However, upgrades to the `pacman` package itself will try to replace the
`pacman.conf` file wholesale. Since the file is modified, a `.pacnew`
file is written instead.

Carefully forward-port the `IgnorePkg` line when replacing the previous
contents with the ones from the `.pacnew` file.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>