sync: avoid running into GitHub's file size limits
GitHub does not allow pushes that include files larger than 100MB. But
LLVM's `libLLVM-16.dll` blasts right through that limit.

Ideally, we would use UPX (https://upx.github.io/) to compress this DLL
file down to size, but UPX is still waiting for Windows/ARM64 support
(see https://github.com/upx/upx/issues/650).

So let's do the next best thing and use `strip.exe` to cut it down to
size.

Sadly, we cannot use the already-available `strip.exe` to do the job
because binutils _also_ is still waiting for Windows/ARM64 support.

To add insult to injury, we cannot even use the already-installed
`llvm-strip.exe` because it is a Windows/ARM64 executable and we're
running the `sync` workflow on hosted runners (which does not include
Windows/ARM64 ones, yet).

So let's bite the apple and install the x86_64 flavor of
`llvm-strip.exe` and use _that_.

This addresses
https://github.com/git-for-windows/git-sdk-arm64/issues/7.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>