Add a PowerShell script to work around those `pacman.exe` hangs
For some reason, there are relatively consistent hangs on Windows/ARM64
when trying to run `pacman.exe` multiple times. Terminating the "right"
processes seems to unblock those hangs (without failure!)

The most likely explanation is a dead-lock in the (x86_64) MSYS2 runtime
where the process that handles signals for child processes is waiting
for that child process, but it is long gone, and the signal handler
process waits forever, blocking its parent process forever.

To help with that, let's implements the PowerShell script outlined in
https://github.com/git-for-windows/git/issues/4883#issuecomment-2045933707
that identifies those hanging processes, waits for them to be idle just
to be extra certain not to disrupt any regular Pacman operation that
waits for a good reason.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>