makepkg-git: do include gettext and error_highlight
In https://github.com/msys2/MSYS2-packages/commit/d7c37ffdadd, `gettext`
was identified as an incorrect dependency of the `xz` package and that
dependency was dropped.

As a consequence, Git for Windows no longer included `gettext`, which
uncovered a bug in the definition of the `makepkg-git` SDK artifact
where it relied on that implicit dependency instead of making it an
explicit requirement (because it is needed by `makepkg` and
`makepkg-mingw`).

Let's make it explicit that `gettext.exe` (and `xgettext.exe`) are to be
included in the `makepkg-git` SDK artifact.

While at it, also avoid the "`error_highlight' was not loaded." warning
when building Git for Windows' `mingw-w64-git-doc-html` package.

This addresses the previously-forgotten x86_64 part of
https://github.com/git-for-windows/git/issues/4952; The i686 part of was
fixed by https://github.com/git-for-windows/build-extra/pull/560
already.

Ported-from-git-sdk-64: a6f190ccbfb (makepkg-git: do include gettext and error_highlight, 2024-07-22)
Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>