makepkg-git: use AsciiDoctor again
In https://github.com/git-for-windows/git-sdk-arm64/pull/6, we switched
the `makepkg-git` SDK artifact in git-sdk-arm64 away from using
AsciiDoctor to using AsciiDoc instead. The reason was that the former
was not yet supported in the `clangarm64` packages in MSYS2.

In the meantime, AsciiDoctor is not only supported, but now also
required to build `mingw-w64-clang-aarch64-git`.

So let's adjust the `makepkg-git` sparse checkout definition
accordingly.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>