makepkg-git: include update-via-pacman.ps1
When trying to work with the build-installers SDK artifact, it turned out that it's missing the update-via-pacman.ps1 file. And since that artifact is using the makepkg-git and minimal-sdk file lists as its foundation, we need to update the file list here, so that said file will end up in the build-installers artifact.

Ref: https://github.com/git-for-windows/git-for-windows-automation/issues/97
Ref: https://github.com/git-for-windows/build-extra/blob/76751073989fa261449a2cc94b8a5defbd8958a5/please.sh#L3307
Signed-off-by: Dennis Ameling <dennis@dennisameling.com>