makepkg-git: include pacman-key and pacman-conf.exe as well
In the previous commit, we added /update-via-pacman.ps1 to the list of files to include in makepkg-git (as this list ends up in the build-installers artifact, which we're trying to use).

However, that script now failed, complaining that pacman-conf.exe and pacman-key are missing.

I confirmed locally that after adding these files manually to the build-installers artifact, the .ps1 file is able to succeed.

Ref: https://github.com/git-for-windows/git-sdk-arm64/commit/fbb8e7344cf8d0ac9fcda4b8a0ea2392c8636ade
Signed-off-by: Dennis Ameling <dennis@dennisameling.com>