build-installers: really adapt to latest Ruby version
To be able to run AsciiDoctor, Git for Windows requires Ruby. MSYS2
recently updated Ruby from v3.1.5 to v3.3.0, which requires updates to
the `build-installers` sparse-checkout definition.

Just like `git-sdk-64` and `git-sdk-32`, we already updated the
definition to include the new `bundled_gems` and `syntax_suggest` files.

But we also need to adapt to the different DLL name in `git-sdk-arm64`.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>