Add a GitHub workflow to downgrade the MSYS2 runtime to v3.5.4
The current version, v3.5.5, suffers from many a problem, two having
seen fixes (a mis-handling of __SIGFLUSHFAST, the other a mis-handling
of the X_OK flag when running as administrator) and at least one more
that causes inexplicable hangs in Git's test suite.

It is too late in the Git for Windows v2.48.0 cycle to investigate these
problems thoroughly and flesh out fixes, therefore the only remaining
option is to downgrade to the last known-okay MSYS2 runtime.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>