ci-artifacts: build Git and run its test suite
Now that Git for Windows no longer needs to rely on Azure VMs in my
personal Azure account to run the Windows/ARM64 automation, it is time
to "un-drop" the more expensive parts of the `ci-artifacts` workflow
that had not been ported from the x86_64 flavor to the aarch64 one.

In fact, this is more than just a revert because the x86_64 flavor has
seen dramatic improvements, e.g. refactoring running the test suite into
a reusable workflow.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>