ci-artifacts: fix running `git version` as part of the build
The `git.exe` that is built needs to have `/clangarm64/bin` in the
`PATH` because it uses a couple of the `.dll` files from that location.

However, the `PATH` does not contain that directory by default. The
reason is that the Bash profile is not evaluated by default by the
`shell:` steps. Let's do that manually, then.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>