libxml2-16: temporarily roll back to allow rebuilding clang
`clang` depends on `libxml2-16.dll`. Turns out that the latter was
rebuilt in https://github.com/msys2/MINGW-packages/commit/2d1af2debd
(libxml2 rebuilds, 2025-09-30) to drop the `DllMain` function.

Turns out that since then, `clang.exe` exits with a 127 (probably
because it somehow assumes that that function exists) and is therefore
unable to rebuild... *drumroll* `clang.exe`.

Let's (temporarily) revert to an earlier state of that DLL so that we
can rebuild the `mingw-w64-llvm` package.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>