ci-artifacts: build git/git again
In 1b7ec02e29a (ci-artifacts: revert to building git-for-windows/git for
now, 2025-04-15), I changed `git-sdk-arm64`'s `ci-artifacts` workflow so
that it builds git-for-windows/git instead of git/git. The reason was
that a lot of work had to be done first before building that branch in
`ci-artifacts` would succeed, which was upstreamed as
https://github.com/gitgitgadget/git/pull/1904 (and which I expected to
take the usual loooong time to trickle down, but Junio seemed to be in a
good mood back then and it took only two weeks and two days).

Since this is all resolved now, let's switch back by reverting
1b7ec02e29a.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>